requirements.txt
emoji_country_flag==1.3.2
pycountry==22.3.5
